+++ apps/demo/src/app/layout/header/notifications-toggle/notifications-toggle.component.ts
@@ -1,13 +1,17 @@
+import { CommonModule } from '@angular/common';
 import { Component } from '@angular/core';
-import { Select, Store } from '@ngxs/store';
+import { MatButtonModule } from '@angular/material/button';
+import { MatIconModule } from '@angular/material/icon';import { Select, Store } from '@ngxs/store';
 import { Observable } from 'rxjs';
 import { NotificationsToggle } from '../../../state/notifications/notifications.actions';
 import { NotificationsState } from '../../../state/notifications/notifications.state';
 
 @Component({
+  standalone: true,
   selector: 'app-notifications-toggle',
   templateUrl: './notifications-toggle.component.html',
   styleUrls: ['./notifications-toggle.component.scss'],
+  imports: [CommonModule, MatButtonModule, MatIconModule],
 })
 export class NotificationsToggleComponent {
   @Select(NotificationsState.enabled)