+++ apps/demo/src/app/markdown/markdown.component.ts
@@ -1,10 +1,13 @@
 import { Component, Input, OnChanges, OnInit, SimpleChanges } from '@angular/core';
+import { DynamicFormMarkdownModule } from '@dynamic-forms/markdown';
 import { MarkdownElement } from './markdown.element';
 
 @Component({
+  standalone: true,
   selector: 'app-markdown',
   templateUrl: './markdown.component.html',
   styleUrls: [ './markdown.component.scss' ],
+  imports: [DynamicFormMarkdownModule],
 })
 export class MarkdownComponent implements OnInit, OnChanges {
   element: MarkdownElement;